Move StudentRecord and related classes to record package in model package
