+++ src/main/java/seedu/taskmaster/logic/LogicManager.java
@@ -15,7 +15,7 @@ import seedu.taskmaster.logic.parser.TaskmasterParser;
 import seedu.taskmaster.logic.parser.exceptions.ParseException;
 import seedu.taskmaster.model.Model;
 import seedu.taskmaster.model.ReadOnlyTaskmaster;
-import seedu.taskmaster.model.session.StudentRecord;
+import seedu.taskmaster.model.record.StudentRecord;
 import seedu.taskmaster.model.student.Student;
 import seedu.taskmaster.storage.Storage;
