+++ src/main/java/seedu/taskmaster/model/ReadOnlyTaskmaster.java
@@ -1,7 +1,7 @@
 package seedu.taskmaster.model;
 
 import javafx.collections.ObservableList;
-import seedu.taskmaster.model.session.StudentRecord;
+import seedu.taskmaster.model.record.StudentRecord;
 import seedu.taskmaster.model.student.Student;
 
 /**
