+++ src/main/java/seedu/taskmaster/model/record/AttendanceType.java
@@ -1,4 +1,4 @@
-package seedu.taskmaster.model.session.record;
+package seedu.taskmaster.model.record;
 
 public enum AttendanceType {
     PRESENT,
