+++ src/main/java/seedu/taskmaster/model/record/ClassParticipation.java
@@ -1,4 +1,4 @@
-package seedu.taskmaster.model.session.record;
+package seedu.taskmaster.model.record;
 
 public class ClassParticipation {
