+++ src/main/java/seedu/taskmaster/model/record/StudentRecordList.java
@@ -1,4 +1,4 @@
-package seedu.taskmaster.model.session.record;
+package seedu.taskmaster.model.record;
 
 import java.util.Iterator;
 import java.util.List;
