+++ src/main/java/seedu/taskmaster/ui/StudentRecordListPanel.java
@@ -8,7 +8,7 @@ import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 import javafx.scene.layout.Region;
 import seedu.taskmaster.commons.core.LogsCenter;
-import seedu.taskmaster.model.session.StudentRecord;
+import seedu.taskmaster.model.record.StudentRecord;
 
 /**
  * Panel containing the list of attendance.
