+++ src/test/java/seedu/taskmaster/logic/commands/MarkCommandTest.java
@@ -12,7 +12,7 @@ import seedu.taskmaster.commons.core.index.Index;
 import seedu.taskmaster.model.Model;
 import seedu.taskmaster.model.ModelManager;
 import seedu.taskmaster.model.UserPrefs;
-import seedu.taskmaster.model.session.AttendanceType;
+import seedu.taskmaster.model.record.AttendanceType;
 import seedu.taskmaster.model.student.Student;
